[YWA-573] Change format of TimeOff start_ and end_time
FE will send us those times in a format without
timezone, for example: '2016-03-22T00:00:00'. All we
want to do, is to add +00:00 at the end to indicate
UTC format with 0 offset.
Also we want to update all existing records. We
change them to UTC format and set database timezone
to UTC.