[YWA-721] Remove working_place validation from employee
Most of the specs are just fixed by not using
first employee working place because that is now
not always true. Employee can exist without
working place. We are explicitly creating
employee working place if needed. If there are
tests that differs between having working place
and not, the additional tests were written to
cover both situations (for example: holidays
vs registered working time).
In update and create event we are taking in
account existence of working place, too.