[YWA-756] Referral program
Create Referrer model and table as
association between users and referred
accounts. Referrer is associated with users
with the same e-mail. It holds token
that is used to refer accounts. Accounts
are associated with Referrer with referred_by
field which is that Referrers token.
Every existing User should already have
a Referrer associated. There's a migration
for that. Every new User triggers Referrer
callback - it checks if Referrer with that
email exists and if not, creates new one.
Account have new param - referred_by, if it's
passed to app, it associates it with Referrer.
There are two new endpoints: one for retrieving
referral token with e-mail - same as for newsletters
and second one for gathering statistics in CSV
file.