[YA-2029] Move class to a separate file and add tests
Changed adjustment_value method also - we need to take into
consideration occupation rate.
Because all the users have been hired on the January 1st, we don't
need to adjust the value accordingly to the number of days till
the end of the year (we can simply take the amount from time off
policy from that time and multiply it by occupation rate of that time).
We could use Adjustments::Calculate and depend on that calculations,
but we don't have employee_time_off_policy connected to the
"hired" type of events for those user (this could be some bug probably).