[YA-2003] Add AccountRemoval class (#598)
* [YA-2003] Add AccountRemoval class

AccountRemoval deletes an account (found by subdomain) and
removes connected resources on Intercom and Stripe.
Unfortunately we're not able to remove intercom company
(there's no supported API endpoint for that), but we can do it
by hand. For intercom we're triggering users removal.
This change will go in one release with #604 where migration
is highly depending on cleaned up database, that's why rake task should
be called before migration.