[YA-1982] Revert assigning time off policies after hiring employee (#612)
Assignations to time off policies are created after the hire event
is created - FE is calling API endpoints for creating assignations
to all of the default time off policies (from default time off
categories like maternity, sickness etc.)
Assignations are therefore created two times. To fix that and move
the whole logic to BE, we need to do proper refactor
and it was agreed to not to do that in the scope of this task
(task for refactor - YA-2091).
After this change - newly hired employee won't be assigned
to time off policies from custom time off categories.
The proper fix will be introduced in the mentioned task.